Tests/LibWeb/Text/expected/close-non-modal-dialog.txt

dialog.open state before close(): true
dialog.open state after close(): false
dialog.returnValue state after close(): after-close
